refactor: new runtime service server suite on runtime_test
