+++ api/handler/v1beta1/job_test.go
@@ -10,80 +10,12 @@ import (
 	"github.com/google/uuid"
 	v1 "github.com/odpf/optimus/api/handler/v1beta1"
 	pb "github.com/odpf/optimus/api/proto/odpf/optimus/core/v1beta1"
-	"github.com/odpf/optimus/core/progress"
-
 	"github.com/odpf/optimus/mock"
 	"github.com/odpf/optimus/models"
 	"github.com/odpf/salt/log"
 	"github.com/stretchr/testify/assert"
 	mock2 "github.com/stretchr/testify/mock"
-	"github.com/stretchr/testify/suite"
 )
-
-type RuntimeServiceServerTestSuite struct {
-	suite.Suite
-	version          string
-	ctx              context.Context //nolint:containedctx
-	namespaceService *mock.NamespaceService
-	projectService   *mock.ProjectService
-	secretService    *mock.SecretService
-	runService       *mock.RunService        // TODO: refactor to service package
-	jobService       *mock.JobService        // TODO: refactor to service package
-	resourceService  models.DatastoreService // TODO: refactor to service package
-	jobEventService  v1.JobEventService      // TODO: refactor to service package
-	adapter          *mock.ProtoAdapter
-	scheduler        models.SchedulerUnit
-	log              log.Logger
-	progressObserver progress.Observer
-
-	req           *pb.DeployJobSpecificationRequest
-	projectSpec   models.ProjectSpec
-	namespaceSpec models.NamespaceSpec
-}
-
-func (s *RuntimeServiceServerTestSuite) SetupTest() {
-	s.version = "v1.0.0"
-	s.ctx = context.Background()
-	s.namespaceService = new(mock.NamespaceService)
-	s.adapter = new(mock.ProtoAdapter)
-	s.jobService = new(mock.JobService)
-	s.log = log.NewNoop()
-	// ... etdc
-
-	s.projectSpec = models.ProjectSpec{}
-	s.projectSpec.Name = "project-a"
-	s.projectSpec.ID = uuid.MustParse("26a0d6a0-13c6-4b30-ae6f-29233df70f31")
-
-	s.namespaceSpec = models.NamespaceSpec{}
-	s.namespaceSpec.Name = "ns1"
-	s.namespaceSpec.ID = uuid.MustParse("ceba7919-e07d-48b4-a4ce-141d79a3b59d")
-
-	s.req = &pb.DeployJobSpecificationRequest{}
-	s.req.ProjectName = s.projectSpec.Name
-	s.req.NamespaceName = s.namespaceSpec.Name
-}
-
-func TestRuntimeServiceServerTestSuite(t *testing.T) {
-	s := new(RuntimeServiceServerTestSuite)
-	suite.Run(t, s)
-}
-
-func (s *RuntimeServiceServerTestSuite) newRuntimeServiceServer() *v1.RuntimeServiceServer {
-	return v1.NewRuntimeServiceServer(
-		s.log,
-		s.version,
-		s.jobService,
-		s.jobEventService,
-		s.resourceService,
-		s.projectService,
-		s.namespaceService,
-		s.secretService,
-		s.adapter,
-		s.progressObserver,
-		s.runService,
-		s.scheduler,
-	)
-}
 
 func (s *RuntimeServiceServerTestSuite) TestDeployJobSpecification_Success_NoJobSpec() {
 	stream := new(mock.DeployJobSpecificationServer)
